Update: Put functions that can be used for Input And Ouput Operations into a base class.
